requirements.txt
flask==3.0.0
google-generativeai==0.7.2
sentence-transformers==2.2.2
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.1
requests==2.31.0
gitpython==3.1.41
kafka-python==2.0.2
streamlit==1.31.0
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.17.0
pdfminer.six==20231228
chromadb==0.4.24
pinecone>=5.0.0
Pillow==10.2.0
onnxruntime==1.18.1
